+++ sqlx-core/src/postgres/mod.rs
@@ -16,6 +16,8 @@ mod error;
 mod executor;
 mod protocol;
 mod row;
+mod sasl;
+mod tls;
 mod types;
 
 /// An alias for [`Pool`][crate::Pool], specialized for **Postgres**.
